python-backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
tensorflow
Pillow
numpy
sqlalchemy
psycopg2-binary
python-dotenv